refactor: Split KanaRomajiMap logic for Server Components compatibility
